+++ src/components/index.js
@@ -6,6 +6,10 @@ export { default as TeacherSideBar } from './sidebar/teacherSidebar';
 export { default as Title } from './textLayout/title';
 export { default as TitleMedium } from './textLayout/titleMedium';
 
+export { default as BottomTwoButtons } from './button/bottomTwoButtons';
+export { default as SubmitButtons } from './button/submitButtons';
+export { default as AddButton } from './button/addButton';
+
 export { default as Teacher } from './layouts/teacher';
 export { default as Director } from './layouts/director';
