+++ src/pages/director/manageCourses/courseDetails.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { Typography, Box, Button, TextField, Grid, Paper, Table, TableHead, TableBody, TableRow, TableCell, TableContainer } from '@mui/material';
-import { TitleMedium } from '../../../components';
+import { Typography, Box, TextField, Grid, Paper, Table, TableHead, TableBody, TableRow, TableCell, TableContainer } from '@mui/material';
+import { TitleMedium, BottomTwoButtons } from '../../../components';
 
 function createData(name, phone, email) {
   return { name, phone, email };
@@ -56,14 +56,7 @@ export default function CourseDetails() {
               </Table>
             </TableContainer>
           </Grid>
-          <Box sx={{ position: 'fixed', bottom: '3vh', right: '3vw' }}>
-            <Button size="large" variant="outlined" sx={{ width: 120, mr: 2 }} onClick={handleCancle}>
-              목록으로
-            </Button>
-            <Button size="large" variant="contained" sx={{ width: 120 }} onClick={handleClickUpdate}>
-              수정하기
-            </Button>
-          </Box>
+          <BottomTwoButtons first="목록으로" second="수정하기" onClickFirst={handleCancle} onClickSecond={handleClickUpdate} />
         </Grid>
       </Box>
     </>
